move language profile to snapshot components
